chore: added an .all-contributorsrc configuration file in the root (#546)
* Create .all-contributorsrc

I created an all-contributors file with all the 10 contributors to the project. All Contributors app on GitHub is an excellent way to acknowledge contributions of the starknet quest project. This will automate the recognition of contributions by adding contributors to the README file of the project.

* Update README.md

Added Markdown for contributor list

* Update .all-contributorsrc

* feat: update with contributors

* feat: remove initial contributors

---------

Co-authored-by: Ayush Tomar <tomar2585@gmail.com>